backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pandas
deepl
requests
pyyaml
